Remove CustomerDataAccess, Add CompanyDataLoader and PersonDataLoader
