+++ java/src/main/java/codingdojo/datawriter/DbDataWriter.java
@@ -1,15 +1,18 @@
 package codingdojo.datawriter;
 
+import codingdojo.dao.CustomerDataLayer;
 import codingdojo.domain.Customer;
-import codingdojo.service.CustomerDataAccess;
+import codingdojo.domain.ImmutableCustomer;
+import codingdojo.domain.ShoppingList;
 import java.util.Collection;
+import java.util.List;
 
 public class DbDataWriter implements DataWriter {
 
-  public final CustomerDataAccess customerDataAccess;
+  private final CustomerDataLayer customerDataLayer;
 
-  public DbDataWriter(CustomerDataAccess customerDataAccess) {
-    this.customerDataAccess = customerDataAccess;
+  public DbDataWriter(CustomerDataLayer customerDataLayer) {
+    this.customerDataLayer = customerDataLayer;
   }
 
   @Override
@@ -24,21 +27,37 @@ public class DbDataWriter implements DataWriter {
   public void write(Collection<Customer> duplicates) {
       duplicates.stream()
         .filter(Customer::isInternal)
-        .forEach(this.customerDataAccess::updateCustomerRecord);
+        .forEach(this.customerDataLayer::updateCustomerRecord);
 
       duplicates.stream()
         .filter(Customer::isNotInternal)
-        .forEach(this.customerDataAccess::createCustomerRecord);
+        .forEach(this.customerDataLayer::createCustomerRecord);
   }
 
   private ACTION updateCustomer(Customer customer) {
-    this.customerDataAccess.updateCustomerRecord(customer);
-    this.customerDataAccess.updateShoppingList(customer, customer.shoppingLists());
+    this.updateCustomerRecord(customer);
+    this.updateShoppingList(customer, customer.shoppingLists());
     return ACTION.UPDATE;
   }
 
-  private ACTION createCustomerRecord(Customer newCustomer) {
-    this.customerDataAccess.createCustomerRecord(newCustomer);
+  private ACTION createCustomerRecord(Customer customer) {
+    Customer newCustomer = ImmutableCustomer
+      .copyOf(customer)
+      .withInternalId("fake internalId");
+    customerDataLayer.createCustomerRecord(newCustomer);
     return ACTION.CREATE;
   }
+
+  public Customer updateCustomerRecord(final Customer customer) {
+    return customerDataLayer.updateCustomerRecord(customer);
+  }
+
+  public Customer updateShoppingList(
+    final Customer customer,
+    final List<ShoppingList> consumerShoppingList) {
+
+    consumerShoppingList.forEach(customerDataLayer::updateShoppingList);
+    return customerDataLayer.updateCustomerRecord(customer);
+  }
+
 }
